pyproject.toml
[tool.poetry]
name = "qdmt"
version = "0.1.0"
description = ""
authors = ["Vinul W"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.2"
scipy = "^1.11.4"
matplotlib = "^3.8.2"
tqdm = "^4.66.1"
ncon = "1.0"
notebook = "^7.0.7"
cssselect = "^1.2.0"
physics-tenpy = "^1.0.0"
quimb = "^1.8.1"
networkx = "^3.3"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
